refactor(auth): replace nextAuthReader with createAuthReaderAdapter\n\n- Introduced createAuthReaderAdapter to streamline authentication handling across various workflows.\n- Removed the deprecated nextAuthReader implementation and updated references in autoResolveIssue, resolveIssueAction, and createDependentPRAction.\n- Added a new Session entity to standardize session management and improve type safety.
