+++ lib/actions/resolveIssue.ts
@@ -1,5 +1,6 @@
 "use server"
 
+import { createAuthReaderAdapter } from "shared/adapters/auth/reader"
 import { makeIssueReaderAdapter } from "shared/adapters/github/IssueReaderAdapter"
 import { EventBusAdapter } from "shared/adapters/ioredis/EventBusAdapter"
 import { OpenAIAdapter } from "shared/adapters/llm/OpenAIAdapter"
@@ -7,7 +8,7 @@ import { makeSettingsReaderAdapter } from "shared/adapters/neo4j/repositories/Se
 import type { GitHubAuthMethod } from "shared/ports/github/issue.reader"
 import { resolveIssue } from "shared/usecases/workflows/resolveIssue"
 
-import { nextAuthReader } from "@/lib/adapters/auth/AuthReader"
+import { auth } from "@/auth"
 import * as userRepo from "@/lib/neo4j/repositories/user"
 
 import { neo4jDs } from "../neo4j"
@@ -69,7 +70,8 @@ export async function resolveIssueAction(
     userRepo: userRepo,
   })
 
-  const authAdapter = nextAuthReader
+  const session = await auth()
+  const authAdapter = createAuthReaderAdapter(session)
 
   const eventBus = redisUrl ? new EventBusAdapter(redisUrl) : undefined
